Split kubectl tests into container vs fluent tests
